Code refactoring. Added put request handling. Added response codes to swagger.

+++ FirstCateringLtd.BackService/Controllers/EmployeeController.cs
@@ -27,34 +27,46 @@ namespace FirstCatering.BackService.Controllers
             _context = context;
         }
 
-        // Returns json array of all employees
-        //GET api/IdCard
+        // GET api/IdCard
+        // Returns array of all employees
         [HttpGet]
+        [ProducesResponseType(typeof(List<Employee>), 200)]
         public IActionResult Get()
         {
             return Ok(_context.Employees.ToList());
         }
 
+
+
+        // Returns an employee from cardId
+
         // Returns 200 Status code with employee data
-        // Returns 400 status code with error message if employee not found
+        // Returns 404 status code with error message if employee not found
         // GET api/IdCard/:id
-        [HttpGet("{id}")]
-        public IActionResult Get(string id)
+        [HttpGet("{cardId}")]
+        [ProducesResponseType(typeof(Employee), 200)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(404)]
+        public IActionResult Get(string cardId)
         {
-            var result = _context.Employees.Find(id);
+            var result = _context.Employees.Find(cardId);
             if(result != null)
             {
                 return Ok(result);
             } else
             {
-                return BadRequest("No employee with this card id");
+                return NotFound("No employee found with this card id");
             }
         }
-
-        // Add an employee with credentials.
+        
+        // Adds an employee with credentials
+        // Returns 200 Status code on completion.
         // Returns 400 Status code if employee with this cardID already exists.
-        // POST api/IdCard
+        // POST api/IdCard/
+        // Body must contain EmployeeNoPin class fields.
         [HttpPost]
+        [ProducesResponseType(200)]
+        [ProducesResponseType(400)]
         public IActionResult Post([FromBody]EmployeeNoPin employeeNoPin)
         {   
             if (!ModelState.IsValid)
@@ -62,19 +74,13 @@ namespace FirstCatering.BackService.Controllers
                 return BadRequest(ModelState);
             }
 
-            if (!cardIdIsValid(employeeNoPin.CardIdNumber)) {
-                return BadRequest("Invalid card id. Card ID can only consist of numbers and letters");
-            }
-
             var existingEmployee = _context.Employees.Find(employeeNoPin.CardIdNumber);
 
             if (existingEmployee != null) return BadRequest("Employee with this card id already exists");
 
-            var cardId = employeeNoPin.CardIdNumber;
+            
 
-            var pin = cardId.Substring(cardId.Length - 4);
-
-            var employee = new Employee(employeeNoPin,pin);
+            var employee = new Employee(employeeNoPin);
 
             _context.Employees.Add(employee);
 
@@ -83,11 +89,46 @@ namespace FirstCatering.BackService.Controllers
             return Ok();
         }
 
-        // Deletes ID card information and employee data by card ID
+        
+        //Replaces the employee with new values
+        //PUT api/IdCard
+        //Body request to have fields of class EmployeeNoPin
+        [HttpPut]
+        [ProducesResponseType(200)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(404)]
+        public IActionResult Put( [FromBody] EmployeeNoPin employeeNoPin)
+        { 
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var employee = _context.Employees.Find(employeeNoPin.CardIdNumber);
+
+            if(employee == null)
+            {
+                return NotFound("No employee found with this card id");
+            }
+
+            employee.Update(employeeNoPin);
+
+            _context.Employees.Update(employee);
+            _context.SaveChanges();
+
+            return Ok();
+        }
+
+        // Deletes ID card information and employee data by card ID        
         // DELETE api/IdCard/:id
-        [HttpDelete("{id}")]
-        public IActionResult Delete(string id) {
-            var employee = _context.Employees.Find(id);
+        [HttpDelete("{cardId}")]
+        [ProducesResponseType(200)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(404)]
+        public IActionResult Delete(string cardId) {
+            var employee = _context.Employees.Find(cardId);
+            
+            if(employee == null) { return NotFound("No employee found with this card id"); }
 
             _context.Employees.Remove(employee);
             _context.SaveChanges();
